requirements.txt
fastapi
uvicorn
python-dotenv
httpx
gunicorn
requests
